ci: automatically generate README
Rather than download and employ fancy tools, we're getting Emacs itself
to generate the HTML. This gives us maximum customisability.

Note that even with the styles turned off, a `<title>` tag is still
produced and this renders strangly on SourceHut. So we remove it with
`sed`, and let the `<h2>` in the `<body>` be the title that renders.